Don't compute 'smothed normals' for positioning off-surface points
The 'smothed normals' were the normals corresponding the the gradient the
target function rather than the normals provided with the input data. Because
the target value at the off-surface points is the value of the target function
there rather than e.g. some constant, the exact position of the off-surface
points doesn't matter too much (more specifically, they may lie on a
self-intersecting line, which would be bad when assigning constant target
values), so we can just as well place them according to the input normals.

Note that this change *does* change the result. The change just shouldn't make
it (much) worse.